refactor(): add service and model-class folders
