+++ src/app/app-container/app-container.component.html
@@ -1,6 +1,2 @@
-<app-content-container [slides]="slides">
-
-</app-content-container>
-<app-drawer-container [slides]="slides">
-
-</app-drawer-container>
+<app-content-container [slides]="slides"></app-content-container>
+<app-drawer-container [slides]="slides"></app-drawer-container>
