+++ src/app/app-container/content-container/content-mask/content-mask.component.css
@@ -2,10 +2,14 @@
   z-index: 9998;
   width: 100%;
   height: 100%;
-  position: fixed;
-  top: 100px;
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
   left: 0;
   display: none;
   background-color: #4c5254;
-  opacity: .7;
+  opacity: 1;
+  transition-property: opacity;
+  transition-duration: 200ms;
 }
