+++ src/app/model-classes/content.ts
