+++ src/app/model-classes/slide.ts
